requirements.txt
annotated-types==0.7.0
asgiref==3.9.0
brotli==1.2.0
cachetools==6.2.2
certifi==2025.6.15
cffi==2.0.0
charset-normalizer==3.4.2
cssselect2==0.8.0
Django==5.2.4
django-admin-adminlte==1.0.14
django-admin-easy==0.8.0
django-admin-soft-dashboard==1.0.22
django-colorfield==0.14.0
django-dbbackup==4.2.1
django-debug-toolbar==4.4.6
django-jazzmin==3.0.1
djangorestframework==3.15.2
fonttools==4.60.1
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-speech==2.34.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
html5lib==1.1
httplib2==0.31.0
idna==3.10
lxml==5.4.0
mysqlclient==2.2.7
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydyf==0.11.0
Pygments==2.19.2
pyparsing==3.2.5
pyphen==0.17.2
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tinycss2==1.5.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.7.0
zopfli==0.4.0
